add sell_shares command
